requirements.txt
django
django-rest-framework
tweepy
django-environ